Compile error for non-ipv6 reported by users
git-svn-id: svn://svn.code.sf.net/p/diameter/svn@208 c7ec899b-d015-0410-9a68-a346fd343c46

+++ cplusplus/libpana/include/pana_ingress.h
@@ -49,7 +49,7 @@ class PANA_EXPORT PANA_IngressJob : public AAA_Job
           m_MsgHandler(0),
           m_Name(name) {
       }
-      virtual int Schedule(AAA_Job* job, unsigned int backlogSize=1) { 
+      virtual int Schedule(AAA_Job* job, size_t backlogSize=1) { 
           return m_Group.Schedule(job);
       }
       virtual int Schedule() {
